Set the input commit on the build to the top-level commit.
When an internal builder runs, currently the src-internal commit is set
on the build. This works fine for trunk since by default the trunk for
src gets checked out and the output gitiles commit gets set based on the
root solution. However, for branched internal builders this fails
because a ref flag needs to be specified for src when the bot_update
script gets called and it doesn't have the ref information for src. This
changes it so that the top-level commit will be used for dependency
projects. Since the input commit now matches the root solution, the
bot_update module will set the necessary --ref flag to get the
appropriate ref of src.

Bug: 1367346
Change-Id: Ia0bd60513874e87fdd8571ebec6ecc65f1a4a9ea
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3916385
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Cr-Commit-Position: refs/heads/main@{#51763}