CTRv1: testlibs -> dynamic port. cros-dut -> host network & dynamicport
Re-enable testlibs on a dynamic port that is found during execution
time.

Change cros-dut to use the host network always. This is required for
local CTR to work with the corp proxy. Additionally, change the
hard-coded port from 80 to dynamic, such that when in the lab, there
isn't port clashes for the default port

BUG=b:249126959
TEST=built locally, ran locally. + staging lab rollout will be needed.

Change-Id: Ia78f52ebf1ec64310400fd7e04f8a22b62c5501a
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3919850
Reviewed-by: Chris DeLaGarza <cdelagarza@google.com>
Reviewed-by: Ming Kong <mingkong@google.com>
Commit-Queue: Derek Beckett <dbeckett@chromium.org>
Reviewed-by: Azizur Rahman <azrahman@google.com>
Cr-Commit-Position: refs/heads/main@{#51792}