Replace build-time paths with a placeholder in sysconfigdata.
These build-time paths should not be used at run time, because they are
not guaranteed to exist. This replaces the build-time install prefix
with a placeholder string "[INSTALL_PREFIX]". We can then replace this
with the actual install location in dockerbuild when we build wheels.

Note that Windows is unaffected by this issue, as it uses a different
code path to populate the sysconfig paths.

Bug: 1367924
Change-Id: Iacefc74f24ba5d5639de3ba526f116329e68b316
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3920730
Reviewed-by: Chenlin Fan <fancl@chromium.org>
Commit-Queue: Brian Ryner <bryner@google.com>
Cr-Commit-Position: refs/heads/main@{#51805}