UFS: Check if pubsub topic exist before creating
Previously, we attempted to create the topic each time a message was
being published. This was leading to a huge amount of errors in GCP and
clouding the logs. This change will check that the topic doesn't exist
before attempting to create the topic.

BUG=b:220373240
TEST=Shivas

Change-Id: Ife6fea5965b81ba33e2e1ce0aa9b08d761ee43d0
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3893228
Reviewed-by: Anushruth Sasivehalli <anushruth@chromium.org>
Commit-Queue: Julio Hurtado <juahurta@google.com>
Cr-Commit-Position: refs/heads/main@{#51983}