[karte] Persist observations
Fix the persist-observations RPC. Write a version of the persist
observations RPC that reads all of the actions in a given range
and then reads all of the observations.

The old RPC was timing out at the ten minute mark, due in part to
the fact that it interleaved attempts to persist observations to
datastore and attempts to persist actions.

Hopefully this will be faster.

If it isn't fast enough, there are a few more tricks we can try,
including:

1) increasing the stride size
2) trying to rewrite the action query to make it more efficient
3) rewriting the RPC to operate on 15-minute chunks instead and use a task queue

BUG=b:236182267
TEST=make test

Change-Id: Ie98583e9ae2743e7bca531b23d9d499a40be89d5
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3935845
Reviewed-by: Xixuan Wu <xixuan@chromium.org>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Cr-Commit-Position: refs/heads/main@{#52026}