[WIB] Add support to upload disk images for online builders (2/2)
Bug: b/243591906

Adds support to upload all the disks that are given by user/config
processor. The disk images that are expected to be used in this are
quite huge and might take a long time on the network to upload. The
images are first compressed(zip) and then uploaded. Not all disk images
will be uploaded, but only those that are specified in the config. The
idea behind this is that the config processor will determine if it
needs to cache any of the images based on parsing all other configs.
If there is some customization that is dependent on one of the disk
images then config processor adds a destination to the config.

Alternatively user can also add a destiniation and that request will be
honored. This can be useful in debugging.

Change-Id: I794f3837d20bc2cc80256fc2b8961291019b9bef