Check out the base revision when bootstrapping an IQR test CL.
In order to get reliable comparisons between full runs and inverted
quick runs, when the bootstrapper is operating on a hard-coded CL that
is used for testing inverted quick runs, the base revision of the CL
will be used instead of getting the tip of tree of the target ref. This
will ensure that the runs have the same version of the test so that
there won't be tests added or removed from one of the runs due to other
changes being made to the test.

Bug: 3942967
Change-Id: I31af890644478244ef7fff9fe63898bccdd2303f
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3945785
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Commit-Queue: Struan Shrimpton <sshrimp@google.com>
Reviewed-by: Struan Shrimpton <sshrimp@google.com>
Cr-Commit-Position: refs/heads/main@{#52118}