Add Read CBI Support to Paris
This code is still not connected to a Critical Action, and is functionally a NO-OP. Ran unit tests for Mallet and Recovery, and then ran against a single Dedede DUT (and manually verified that CBI contents were read) using a custom plan that only references CBI actions: https://paste.googleplex.com/6609060378968064. Broader testing will be conducted before connecting this change to a critical action.

See go/fleet-recovery-developer for more information on how to build and test Paris changes in general.

BUG=b/235000813
TEST=manual

Change-Id: Iba316f9bc04a827f768b30b4c355a6dadb898905