[recovery]: Fixed the comparison of error value from RW VPD list CMD.
The error code in the runner is written into an empty interface, and
is of type int32. Later, when we check the value of error code, we
compare it with plain int literals, and this comparison fails even if
the error values are the same.

This commit fills this gap by casting the int literals to int32, which
can then be compared to the value stored in the interface.

BUG=b:253326688
TEST=build, go test ./..., manual testing

Change-Id: I92250d388f494eb406444cd1043f3e51ffe7e06c
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3961946
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Vineet Joshi <vkjoshi@google.com>
Cr-Commit-Position: refs/heads/main@{#52282}