Remove abennetts@google.com from appengine/chromium_status_py3/OWNERS.
This suggested change is automatically generated based on group memberships and affiliations.
If this change is unnecessary or in error, vote CR -1 and the bot will abandon it. Vote CR +1/2 to approve this change.
See the owner's recent activity for context: https://chromium-review.googlesource.com/q/abennetts@google.com

To report an issue, file a bug in the Infra>Security component.

Change-Id: I2f9ed6b79c3774ca3561fa210e4932de0a92b017