Remove abennetts@google.com from go/src/infra/appengine/sheriff-o-matic/OWNERS.
This suggested change is automatically generated based on group memberships and affiliations.
If this change is unnecessary or in error, vote CR -1 and the bot will abandon it. Vote CR +1/2 to approve this change.
See the owner's recent activity for context: https://chromium-review.googlesource.com/q/abennetts@google.com

To report an issue, file a bug in the Infra>Security component.

Change-Id: Ib4518001e6d1df6c1d7395c9c055bf4466a341e6