Remove abennetts@google.com from go/src/infra/appengine/test-results/OWNERS.
This suggested change is automatically generated based on group memberships and affiliations.
If this change is unnecessary or in error, vote CR -1 and the bot will abandon it. Vote CR +1/2 to approve this change.
See the owner's recent activity for context: https://chromium-review.googlesource.com/q/abennetts@google.com

To report an issue, file a bug in the Infra>Security component.

Change-Id: I1c36b91fed401105e64bf4c8725eb0408f9c9e92