RTS: don't use origin/main as reference
The problem doesn't seem completely resolved. The only difference
remaining is the lack of a ref. See https://crrev.com/c/3972735
for tests verifying this still works. The only reason to have a
change ref would be for local devs to compare to another branch so
I'm adding an arg just in case anyone wants to run this locally

Bug: 1361482
Change-Id: I92e9ee006378e74b8f32aba500ace50c0e1d2490
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3973590
Reviewed-by: Gregory Guterman <guterman@google.com>
Commit-Queue: Struan Shrimpton <sshrimp@google.com>
Cr-Commit-Position: refs/heads/main@{#52419}