CTR: fix CTR provision errs from being swallowed, + incr poll timeout
There is a KI in fleet where the disk i/o is maxed out, causing services
to take serveral minutes to start on occasion, it was recommended to
increase this to 3 minutes for now. Additionally fixed the logs so
rather than a silent exit we can see the err msg here.

BUG=b:255782067
TEST=build,then local run with the failure to see the logs.

Change-Id: I619217133e4e669549bc3fcc439eee48ddef897c
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3984114
Reviewed-by: Azizur Rahman <azrahman@google.com>
Commit-Queue: Derek Beckett <dbeckett@chromium.org>
Auto-Submit: Derek Beckett <dbeckett@chromium.org>
Cr-Commit-Position: refs/heads/main@{#52483}