Adds Exec To Restore CBI Contents From UFS
Adds an exec to restore previously backed up CBI contents from UFS to the DUT. This action is functionality a NO-OP for the time being as there is no critical action connecting it to the repair flow. At the moment, this has only been tested on a single DUT. Broader testing will be conducted, similar to the testing for backup cbi before we connect this change to a critical action.

Test Artifacts: https://paste.googleplex.com/4797048875384832

BUG=b/235000813
TEST=Manual and unit tests

Change-Id: I215910cf7b47202dee78ac3d220e6a5594cbfb89
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4010642
Commit-Queue: Sean Carpenter <seancarpenter@google.com>
Reviewed-by: Otabek Kasimov <otabek@google.com>
Cr-Commit-Position: refs/heads/main@{#52788}