Adds an Action To Check CBI For Corruption
Adds a check to detect CBI corruption and enables the check in production by adding it as a critical action. After verifying that the check works in production, we'll later add a recovery action with the repair CBI logic to it.

This test was changed by running it against DUTs in the Santiam lab (a combined 20 unique boards).

Test Artifacts: https://b.corp.google.com/issues/235000813

BUG=b/235000813
TEST=Manual and unit tests

Change-Id: I73e0fcd9bad17be57e6149c14b3106dac9556050
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4018325
Commit-Queue: Sean Carpenter <seancarpenter@google.com>
Reviewed-by: Otabek Kasimov <otabek@google.com>
Cr-Commit-Position: refs/heads/main@{#52810}