Adds an Action To Check For CBI Magic Validity
All healthy instances of CBI Contents start with the bytes "CBI". This check already existed in the logic for
backing up CBI contents, but is now made explicit to make it easier to query for in Karte. This is the one
check that we can do that will give us absolute certainty that CBI is corrupt on the DUT, so good to have.

Test Artifacts: https://b.corp.google.com/issues/235000813#comment43

BUG=b/235000813
TEST=Manual and unit tests

Change-Id: I3baaef3d1454988086b26df665604b77b251a3a4
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4031575
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Sean Carpenter <seancarpenter@google.com>
Cr-Commit-Position: refs/heads/main@{#52985}