Ensure that no two wheel builds will attempt to output to the same file.
This fixes a race condition which can happen if a wheel is being built
as a dependency of one or more wheels. The dependent build might access
the output wheel file as it is being overwritten by a build on another
thread, resulting in pip seeing a truncated, invalid wheel file. Fix
this by using an output directory under the temporary build directory.

Also ensure that no two top-level wheels will attempt to output to the
same file, which could result in a similar race.

This is a quick fix to work around the lack of an actual dependency
graph. Future improvements would use the more optimal behavior of only
building the dependency once.

Also make the pip cache per-thread, as sharing the cache seems to lead
to intermittent errors when installing wheels from pypi.org. The pip
cache also now lives under the build root, rather than in the user home
directory.

Bug: 1384388
Change-Id: Idbf4f756222ab23ad5bd800950a0496e955d6f13
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4036472
Reviewed-by: Chenlin Fan <fancl@chromium.org>
Commit-Queue: Brian Ryner <bryner@google.com>
Cr-Commit-Position: refs/heads/main@{#53045}