[WIB] Add exists api to sources.py
Bug: b/204104098

Adds exists and updates get_url api. This will allow the recipe to
determine which packages/artifacts/files exist at remote and make
decisions to the order of execution of various customizations. These are
not covered by tests as they will be used in scheduler and as a result
covered by the scheduler tests

Change-Id: I884efebc8cd944a0859e20513e3c32bfe58bd28b
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4010640
Commit-Queue: Anushruth Sasivehalli <anushruth@chromium.org>
Reviewed-by: Vineet Joshi <vkjoshi@google.com>
Cr-Commit-Position: refs/heads/main@{#53173}