sched-labels: add method to expose labels from HWID server
The original design was to expose labels using Boxster configs. But as
the HWID server is the source of truth of HWID component information,
UFS can expose all HWID component labels by using HwidData. This
implementation exposes all HWID components for a HWID design as
schedulable labels.

A HWID component is stored in the form `xxx/yyy`. The label will be
exposed as `hw-xxx: yyy`.

BUG=b/186663540
TEST=make test

Change-Id: I138b1007f04c23c03140b6289d6a0bb4b0b26fe3