[🚊] Remove simple_search_stub.py and dependencies from third_party/.
simple_search_stub.py is a simple RAM-backed Search API stub, probably
primarily for testing and local development. It doesn't make RPC calls
to the actual Search API. We already mock out Search calls in unit
tests, so we don't need it. It's easier if we exclude it, because it
requires some additional work to make Python 3-compatible.

Bug: monorail:11600
Change-Id: I28f066be062c6ff6fe90c0b76fef312483e1ed9a