[WIB] Redesign scheduler to package tasks
Bug: b/204104098

This change if applied redesigns scheduler to package customizations
more effectively. The `Image Config Processor` will package tasks that
can be run on a single builder into an image config and schedules them.
The processor also waits for the builds to complete and schedules
remaninig builds (if any). This is a major change from the old behavior
which triggered all tasks (customizations) individually and any task
that was dependent on another would fail.

Change-Id: If8e79d8c4613ae96727d2c40c81210e348b17875
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4009484
Commit-Queue: Adam Todd <actodd@google.com>
Reviewed-by: Adam Todd <actodd@google.com>
Cr-Commit-Position: refs/heads/main@{#53250}