debug symbols: properly handle duplicate file names
See bug for full context. The gist is this:
There can be multiple symbol files with the same base name. If the
files contain different symbol IDs, we want to upload both (and
need to add a unique suffix to the local copy of files to avoid
duping / overwriting). If the files contain the same symbol ID,
they are legitimate duplicates and should not be uploaded multiple
times.

BUG=b:261929316
TEST=go test

Change-Id: I2dc3ed0cef2fd83965b4bbf91998edebb51d9ae6
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4097744
Commit-Queue: Jack Neus <jackneus@google.com>
Reviewed-by: Julio Hurtado <juahurta@google.com>
Cr-Commit-Position: refs/heads/main@{#53472}