otil: Delay tracer construction
The original code followed an example, but tracer construction happens
immediately and is not lazy, so we need to delay it until after the
TracerProvider+exporter is set in main.  Otherwise, the traces go to
the default (no-op) exporter.

Other examples also use this pattern and it should be sufficiently lightweight.

BUG=b:259104564
TEST=go test

Change-Id: I6270664534100c0de785eb8cbbcaa14ff2c1a7ab
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4107772
Reviewed-by: Congbin Guo <guocb@chromium.org>
Commit-Queue: Allen Li <ayatane@chromium.org>
Cr-Commit-Position: refs/heads/main@{#53542}