test_plan: Use MappingForm_ORIGINAL in validate command.
- There are cases where the regex check will fail when
inherited in a subdirectory. For example, if there is
a regex "a/b" in a DIR_METADATA in subdirectory "a",
and there is a DIR_METADATA in subdirectory "a/c",
the inherited regex ("a/b") will not be a prefix of the
subdirectory ("a/c"). Use ORIGINAL mapping, which doesn't
include inherited properties, when validating.

- Also sort directories and report all errors.

BUG=b:241975591
TEST=CQ
TEST=`test_plan validate ~/chromiumos/src/platform2`

Change-Id: I184eb50c583fe8aee897ff2178f1b632f65aaa35
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4108229
Auto-Submit: Andrew Lamb <andrewlamb@chromium.org>
Commit-Queue: Navil Perez <navil@google.com>
Reviewed-by: Navil Perez <navil@google.com>
Cr-Commit-Position: refs/heads/main@{#53544}