Update the gitiles client to support getting revision for a path.
In order to enable bootstrapped builders to be able to skip checking
out, we need to be able to ensure that the version of the source checked
out onto disk is compatible with the properties that the bootstrapper
loads. To do this, we need to know the last commit that modified the
properties file.

Bug: 1396163
Change-Id: Icdb4007787a11f129f18d502482832dc58d09dff
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4108958
Reviewed-by: Struan Shrimpton <sshrimp@google.com>
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Commit-Queue: Struan Shrimpton <sshrimp@google.com>
Cr-Commit-Position: refs/heads/main@{#53552}