Communicate the last changed config commit to bootstrapped builds.
After downloading the properties file, the bootstrapper will issue
another request to gitiles to find out the last revision that the
properties file was changed at. This information along with the path to
the properties file will be communicated to the bootstrapped build so
that it can determine whether it needs the source on disk to be updated
or not.

Bug: 1396163
Change-Id: I5c252966579e05a5453a5f6dee8994869eb898c9
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4108910
Reviewed-by: Struan Shrimpton <sshrimp@google.com>
Commit-Queue: Garrett Beaty <gbeaty@google.com>
Cr-Commit-Position: refs/heads/main@{#53572}