[recovery]: Updated the destination of the logs to a new folder.
In Paris, we collect logs before we start a repair process. If the DUT
is not reachable over ssh, we attempt to collect the logs as soon as
the DUT becomes reachable, provided the logs haven't been lost.

During this log collection, we want to place the logs within a folder
called "prior_logs". This commit updates the name of the
destination folder for this.

Bug=b:235266300
TEST=build, go test ./...

Change-Id: Id6d5b127e43b833c8ea6b8dbc7367e3cd3cdb6c2
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4130926
Commit-Queue: Vineet Joshi <vkjoshi@google.com>
Reviewed-by: Otabek Kasimov <otabek@google.com>
Cr-Commit-Position: refs/heads/main@{#53751}