Remove myself from 3pp owners.
It seems unlikely that we really need me in here for emergency approvals
at this point; Worst case we could use Owners-Override. It would be nice
if there was syntax for this in OWNERS to allow someone who can approve
but is never selected automatically by the tools.

Change-Id: I91356d3f0c67d3f38a4c2c9a0ac5b04bd4be39d2
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4144886
Reviewed-by: Brian Ryner <bryner@google.com>
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Cr-Commit-Position: refs/heads/main@{#53853}