Rename dockerbuild platforms for clearer naming going forward.
Platforms ending in 'py3', which are really specific to 3.8, are renamed
to end in 'py3.8'. Mac platforms, which were the only ones ending in
'cp38', are renamed to 'py3.8' as well.

This CL adds the new platform names everywhere that the old names
are used, but does not remove the old names. That will happen after
we have switched the builders to pass in the new names.

Add a field dockerbuild_image that allows specifying the name of the
docker image, which no longer has to match the platform name. This
also allows us to add variants of our Linux platforms without
rebuilding the Docker images each time.

Bug: 1399691
Change-Id: I85ca662189df954e2400a7720f8102876877f9a6
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4167875
Reviewed-by: Chenlin Fan <fancl@chromium.org>
Commit-Queue: Brian Ryner <bryner@google.com>
Cr-Commit-Position: refs/heads/main@{#53998}