Update the bootstrapper to be more tolerant of DEPS changes.
Currently, if a builder using a dependent repo bootstrap configuration
is run against a CL for the top-level repo that modifies the DEPS file
then it attempts to apply the DEPS patch onto HEAD of the DEPS file in
order to get the pin for the dependent repo in a way that works whether
the DEPS file modifies the desired pin or not. However, this causes
problems if between the CL's base revision and HEAD there were changes
to lines near lines modified by the CL because it prevents the patch
from being able to apply. This is not a problem for bot_update because
it ensures a fill git repo checkout and can rely on using git
cherry-pick.

This change makes the bootstrapper resilient to that situation by not
trying to apply the DEPS patch. Because the bootstrapper is only
interested in the value of the pin for a single dependency, it doesn't
need a fully consistent patched file and can just compare the value of
the pin at different revisions in order to decide what pin to use. By
comparing the revision of the DEPS file in the CL with that of the DEPS
file in the base revision, we can determine if the CL changes the value
of the dependency pin. If it does, then use the new revision specified
in the CL. If the pin is unchanged, then we get the revision from the
head version of the DEPS file since that is what the builder will apply
the CL on top of.

Bug: 1408500
Change-Id: I608fe5d6ab5591621ad8a72deb1890881bdb02c5
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4179314
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Commit-Queue: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Cr-Commit-Position: refs/heads/main@{#54029}