Only skip setting build summary for bootstrapped exe failure.
Currently the bootstrapper's main code looks for an exec.ExitError which
indicates that a process was launched and that it returned a non-0 exit
code. The intention is to detect that the recipe was successfully
triggered, in which case the bootstrapper would have ceded control over
the build to the recipe. However, it is incorrectly skipping updating
the build for all exec.ExitErrors, including those from processes that
the bootstrapper uses to perform its own operations (cipd, gclient, git
patch). This means failures in those cases show up as an infra failure
with no information.

This change modifies the code so that the ExitError detection is
performed only after the call to execute the bootstrapped executable.
This will result in failures in other executables producing a more
helpful build summary.

Change-Id: I663127c0494f4a80a523ab92fe17e1e2ad73d1ed
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4192053
Reviewed-by: Struan Shrimpton <sshrimp@google.com>
Commit-Queue: Garrett Beaty <gbeaty@google.com>
Cr-Commit-Position: refs/heads/main@{#54122}