phosphorus: Short circuit firmware provisioning
... from what firmware labels are desired and found.

There are cases of rw firmware being provisioned when it's not required.
The special case is when ro is targeting the same version as rw.

Previously the provisioning when sent both ro and rw labels that were
targeting the same versions, ran FirmwareUpdate and FirmwareUpdate.rw_only
in sequence.

BUG=b:266469356
TEST=go test

Change-Id: I5b68535c18ebbc48c412bd528f16f7cd3971555a
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4206499
Reviewed-by: Xixuan Wu <xixuan@google.com>
Commit-Queue: Jae Hoon Kim <kimjae@chromium.org>
Cr-Commit-Position: refs/heads/main@{#54289}