testplan: Only clone up to a depth when patching CLs.
- Repos that are large / have large histories can have long
clone / fetch times. Only clone up to a depth (past the earliest
CL in a repo); set 180d as the default depth. As long as CLs don't
have a parent that is very old they should be able to merge (and
can fall back to cherry-picking otherwise).

BUG=b:266220792
TEST=CQ
TEST=Run on
https://chromium-review.googlesource.com/c/chromiumos/third_party/linux-firmware/+/3880708.

Change-Id: I8104a382926a842c39877bc014b8ddf082717c0a
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4199710
Reviewed-by: Navil Perez <navil@google.com>
Commit-Queue: Andrew Lamb <andrewlamb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#54295}