[WIB] Add parted 3.5 support to create disks
Add support to create disks using parted. This ensures that we can
create disks without any special permissions.

A straightforward way to create disks is to create an image, mount it
and then partition, format ... as needed. This means we need a lot more
permissions to be able to access disks. A workaround to this involves
creating seperate image for each partition, formating the image using
`mkfs` and then reassembling the disk image. We still need to write a
partition table to this otherwise it won't be readable. Luckily, parted
has a rescue command that can repair disks by updating the partition
table. We need parted 3.5 becuase bionic shipped with 3.2 and it comes
with a bug where you need user input to confirm the changes even when
`-f` force is used.

Change-Id: I8fcf32c24d4cfe07b88ab6de0246fa1993ab6603
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4144885
Commit-Queue: Anushruth Sasivehalli <anushruth@chromium.org>
Reviewed-by: Adam Todd <actodd@google.com>
Cr-Commit-Position: refs/heads/main@{#54306}