[lkgr-finder] Speed up process with partial checkouts
The lkgr-finder service has its own checkout of the supported project
in order to sort revisions and extract commit positions. Full naive
checkouts are very slow in the Chromium case.

This CL makes the internal checkout partial by omitting trees. They
are not needed for commit history and commit message parsing. The
resulting speed-up when running the script with a fresh workdir
is >x5.

When this is rolled out, a `git fetch` omitting trees might be
executed on an existing workdir that already fetched trees, which
is no problem. Even if a command that requires trees like `show`
or `checkout` was performed, it'd lazy-fetch the required trees.

This CL removes an unused feature that'd allow to also set a tag.
The recipe doesn't use this, and the only consumer of this script
is the recipe.

Usage of `git show -s` are turned into `git log -1`, which is
equivalent in the output, but the former fetches the trees
inefficiently, just to filter them out later.

The git library has no tests unfortunately. This was tested by
executing the script locally.

Bug: chromium:1412804
Change-Id: Iaf37ef08db9fb72dcc8af629fdfcb43ad3d507ef
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4224674
Commit-Queue: Michael Achenbach <machenbach@chromium.org>
Reviewed-by: Alexander Schulze <alexschulze@chromium.org>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Cr-Commit-Position: refs/heads/main@{#54464}