recovery: Add plan run tally as observation to metrics
Make it easier to find the last execution of a particular action in a plan.
Remove the forgiven tally; it is redundant since we already store whether actions are allow-fail or not.
Make the execution time more detailed; store it as seconds.

BUG=None
TEST=build, run mallet

Change-Id: I3ba44d19bb988c8258f994b1996fdf8dbaf178d3
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4237152
Commit-Queue: Otabek Kasimov <otabek@google.com>
Reviewed-by: Gregory Nisbet <gregorynisbet@google.com>
Auto-Submit: Otabek Kasimov <otabek@google.com>
Cr-Commit-Position: refs/heads/main@{#54557}