[golangbuild] propagate build status correctly
Currently, the returned error in a few functions doesn't properly attach
the build status on the returned error, only the error passed to
step.End. This results in a bad UX wherein a build looks like it failed
normally, but when you drill down you find the failing step was actually
an infra failure.

Change-Id: Ica2724a4d3e2e7a2e5235c1f7025a604fd966bde
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4246857
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Commit-Queue: Michael Knyszek <mknyszek@google.com>
Cr-Commit-Position: refs/heads/main@{#54644}