Don't set a top level input commit when bootstrapping a dependency CL.
When bootstrapping a CL for the dependency project, the bootstrapper
doesn't download anything from the top-level repo. Currently, it is
setting an input commit for the configured top level ref. This results
in an infra failure in the bootstrapped recipe if the builder isn't
checking out the top level repo (e.g chrome/try/branch-config-verifier
only checks out src-internal) because bot_update tries to match the repo
of the commit against the paths checked out in the gclient config. This
changes it so that the commit for the top level repo isn't set so that
such builders can pass bot_update without having to add unnecessary
configuration.

Bug: 1414475
Change-Id: I8222ae04a14dd1b26f76392c4f5b6b24802da8d1
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4246377
Reviewed-by: Struan Shrimpton <sshrimp@google.com>
Commit-Queue: Garrett Beaty <gbeaty@google.com>
Cr-Commit-Position: refs/heads/main@{#54670}