🚝 Add third_party/protorpc/remote.py.
protorpc generally supports Python 3, except that protorpc.remote
uses `async` as a field name, which is a reserved keyword starting
in Python 3.5, with the introduction of coroutines. protorpc has
been unmaintained and archived for years, and will not take pull
requests. So, we have a local copy of that file with `async` renamed
to `async_`, and modify the package __path__ to use our local copy.

Bug: monorail:5271
Change-Id: I7e8ee3f6fe096e62ff0176d1fc940c61c6df8bbf
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4257712
Reviewed-by: Andrew Chang <andrewjc@google.com>
Commit-Queue: Andrew Chang <andrewjc@google.com>
Auto-Submit: Dave Tu <dtu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#54736}