🚝 Python 3: Fix `TypeError: unhashable type: 'Issue'` etc.
In Python 3, "A class that overrides __eq__() and does not define __hash__() will have its __hash__() implicitly set to None." This is
new behavior that was not true in Python 2.

protorpc.messages.Message falls under this category. Therefore, its
descendants (ComponentDef, Issue, User, etc.) are unhashable and
cannot be added to sets or dict keys.

The hashing behavior in Python 2 was undefined, so I broadly think
it's safe to just change all the sets to lists, as the deduplication
probably didn't work consistently anyway.

Alternative considered was to add a __hash__() function to
protorpc.messages.Message. However, it's a mutable type, so that
would also lead to undefined behavior. "If a class defines mutable objects and implements an __eq__() method, it should not implement
__hash__(), since the implementation of hashable collections requires that a key’s hash value is immutable."

Bug: monorail:5271
Change-Id: I0f3f723e76c47fc082c9560df4ac94036978eb01
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4288284
Commit-Queue: Dave Tu <dtu@chromium.org>
Auto-Submit: Dave Tu <dtu@chromium.org>
Reviewed-by: Yuxuan Chen <yuxuanch@google.com>
Cr-Commit-Position: refs/heads/main@{#54897}