[package_index_cros] Support stable packages.
Before the change the script only worked with work-on packages and
9999 ebuilds. The package's paths were hardcoded to look for 9999
version.

Now the script is able to work with packages of stable revision and
processed packages are no longer required to be worked on.

This should help to decouple the build step from the script.

Bug: b:269132220
Test: manual
Change-Id: I72d96dcf9a6b3355172a3525a96e4e3af331be95
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4308399
Commit-Queue: Artem Sumaneev <asumaneev@google.com>
Reviewed-by: Gavin Mak <gavinmak@google.com>
Cr-Commit-Position: refs/heads/main@{#55199}