Update dockcross for linux-arm platforms to newer, LTS images.
These images have an old glibc (2.27) but a newer gcc (8) which
should be a reasonable combination for our supported platforms
and packages, for now. Rewrite the python sysconfig to account
for some renaming that has happened, and fix a bug in locating
the sysconfig files when running in docker.

Also update the manylinux-x64 image to match, though it is now
on gcc 10.

Bug: 1422445
Change-Id: Id4ea96e928122851385f09cb60ed02a1565a8aa6
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4318543
Reviewed-by: Chenlin Fan <fancl@chromium.org>
Commit-Queue: Brian Ryner <bryner@google.com>
Cr-Commit-Position: refs/heads/main@{#55351}