🚝 Port in fix for google/appengine/runtime/default_api_stub.py.
In appengine-python-standard==1.0.0, default_api_stub throws an error
when trying to access NDB outside of a Flask request. This was fixed
in commit cc19a2e on Juy 21, 2022, but wasn't included in the 1.0.1rc1
release on Sep 6, 2022. It's been months since that release, so instead
of waiting on another release, we'll just monkeypatch the file here.

Bug: monorail:5271
Change-Id: I377865b592a0ad4a0a881733edcf02ca571cfee0
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4348897
Reviewed-by: Andrew Chang <andrewjc@google.com>
Commit-Queue: Dave Tu <dtu@chromium.org>
Reviewed-by: Yuxuan Chen <yuxuanch@google.com>
Cr-Commit-Position: refs/heads/main@{#55494}