🚝 Port in fix for google/appengine/runtime/default_api_stub.py.
Reland of https://crrev.com/c/4348897 with `if six.PY3:`

In appengine-python-standard==1.0.0, default_api_stub throws an error
when trying to access NDB outside of a Flask request. This was fixed
in commit cc19a2e on Juy 21, 2022, but wasn't included in the 1.0.1rc1
release on Sep 6, 2022. It's been months since that release, so instead
of waiting on another release, we'll just monkeypatch the file here.

Bug: monorail:5271
Change-Id: Ia78af4810e932c980f9dac0c7ed61ba204315b6b
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4354774
Commit-Queue: Andrew Chang <andrewjc@google.com>
Auto-Submit: Dave Tu <dtu@chromium.org>
Reviewed-by: Andrew Chang <andrewjc@google.com>
Cr-Commit-Position: refs/heads/main@{#55538}