vm-leaser: implement releaseExpiredVMs cron job
This cron job utilizes the GCE REST API to list and delete expired VM
instances based on the expiration_time metadata. The REST API currently
has a bug that prevents a direct filtering using the metadata items. As
a workaround, this implementation iterates through the list of running
instances and checks the expiry for deletion. The deletion step takes
about 2 minutes to complete so this cron job will be set to every 5
minutes for now.

BUG=b/271013310
TEST=make build-cron; make dev-cron

Change-Id: I9c7bf30854644703b606a49d8ed8c99b61489cea
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4339281
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Deniz Kara <denizkara@google.com>
Commit-Queue: Justin Suen <justinsuen@google.com>
Cr-Commit-Position: refs/heads/main@{#55543}