Bundle swarm_docker and docker_devices into a single vpython-only pkg.
This removes the ENV dependency from this cipd package, and makes the
resulting package platform-agnostic (it says linux-amd64 so that only
one of the builders will try to package it).

The new package is called infra/swarm_docker_tools, and includes
android_docker, cros_docker and swarm_docker (plus all their
infra-repo dependencies).

We will need to roll this change out in puppet as well.

I did very basic verification that the bundle operates correctly using
vpython_main the same way it appears to be invoked from swarming (i.e.
the script seems to get to argumenting parsing in `main`, which implies
that all importable things worked). I also did a simple static analysis
that all imports are either contained in the embedded vpython spec or
are part of the CIPD package.

This is being done in preparation for removal of the ENV stuff from
infra.

R=bpastene, bryner, vadimsh

Bug: 977627, 1413668
Change-Id: I1f03debdfb30aafd25aec95ca29ffc73595f46a8
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4358830
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Cr-Commit-Position: refs/heads/main@{#55565}