Uprev CTP library
Planned to migrate the impact of this CL: https://crrev.com/c/4348320,
which requires a BBClient to be explicitly set, but crosfleet provides
it's own client explicitly, so there is no need.

Still, figured it's good practice to use the updated dependency in the
off chance the explicit BB client is removed.

BUG=b/272271336
Test Plan=unit tests, compiling and running crosfleet

Change-Id: Iab04c92eded7a3d57ce36cc84f24aa397631b52e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4361923
Reviewed-by: Jard Loucks <jaredloucks@google.com>
Commit-Queue: Eli Trexler <elijahtrexler@google.com>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#55616}