recovery: Make sample sleep always pass and skip metrics
Sleeping action sometimes fail as do not have enough resource to perform an action in time. The sleeping action only design to hold an execution, and if it fail it still succeed the task, so setting it to successful state is what we need.

BUG=None
TEST=build

Change-Id: Ib661cb732d9c3d596abff334bac5d5fff224cf28
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4382337
Auto-Submit: Otabek Kasimov <otabek@google.com>
Reviewed-by: Gowri Denduluri <gowriden@google.com>
Commit-Queue: Gowri Denduluri <gowriden@google.com>
Cr-Commit-Position: refs/heads/main@{#55766}