[cpython3] Remove availability of crypt module.
This should skip the inclusion of `crypt` which is already deprecated
in python3.11 and has extremely spotty support depending on which system
the module is compiled on.

Note that the `crypt` module is NOT modern cryptography, but is instead
insecure primitives related to the generation of /etc/passwd hashes,
such as DES, MD5, etc.

Cursory code search does not reveal any imports of this module.

R=bryner

Bug: 1310140
Change-Id: I4939896f93f3402c5ea91f7eea2f90834cfa69eb
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/4379120
Reviewed-by: Brian Ryner <bryner@google.com>
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Cr-Commit-Position: refs/heads/main@{#55781}