[result_adapter] preserve original test event order for test records
This change is a refactor that changes the order in which test records
are returned, but not their content.

goRun.generateTestResults currently parses incoming 'go test -json'
test events in goTestJsonToTestRecords, and passes only a map keyed
by test IDs to testRecordsToTestProtos. That means test records are
produced in a non-deterministic order.

It seems there's no benefit to emitting test records in an unordered
fashion. More immediately, it makes adding a test case for more than
one test record hard. Though it's also possible to sort test records
by some field (perhaps start time), the simpler solution is to use a
slice that preserves the incoming order.

Change-Id: Ida725ae614128a57e99882ab1d7fefc7cc68e581