[result_adapter] make test output printing more intentional
Previously, goTestJsonToTestRecords unconditionally printed test output
to stdout. This can be quite confusing when doing development on it and
when running its tests. It's also easy to mistake for an unintentional
print statement left over from debugging rather than an intentional
behavior added and motivated in crrev.com/c/3133951.

Refactor it behind a new goRun field and enable only when result_adapter
runs as a command. Continue to use output fragments in the same order as
the test events.

Change-Id: If9f703a63626d7f7738e8859e6cf31e70cd1832f